Add logging of various items.
Items that can be logged:
timeStamp: the current time at which the log was called.
line     : the current line at which the log function was called from.
fileName : the current file at which the log function is called from.
threadID : the current thread that called log.
processID: the id of the process which called log.